Allocated BruteForce to a class and changed the structure of the app.
